requirements.txt
pandas
numpy
flask
mysql.connector
gunicorn
datetime
PyJWT==v1.7.1







